Move VectorMapRenderUtils to the correct package, we no longer need to access geotools package private fields
